# game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/game-api/game-scores/{id}": {
      "get": {
        "operationId": "GameScoreController_getById",
        "summary": "Get game score by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GameScores"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "GameScoreController_update",
        "summary": "Update game score by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameScoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GameScores"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "GameScoreController_delete",
        "summary": "Delete game score by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GameScores"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/game-api/game-scores": {
      "post": {
        "operationId": "GameScoreController_create",
        "summary": "Create game score",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameScoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameScoreResponse"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GameScores"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/game-api/games/{id}": {
      "get": {
        "operationId": "GameController_getById",
        "summary": "Get game by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "GameController_update",
        "summary": "Update game by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "GameController_delete",
        "summary": "Delete game by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/game-api/games": {
      "post": {
        "operationId": "GameController_create",
        "summary": "Create game",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Board Game API",
    "description": "Auth, admin, and game domain API for the board-game-app",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "example": [
              "email must be an email",
              "password should not be empty"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "path": {
            "type": "string",
            "description": "Request path; included by some global exception filters",
            "example": "/auth/login"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp; included by some global exception filters",
            "example": "2026-08-01T07:30:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "HttpErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "GameScoreResponse": {
        "type": "object",
        "properties": {}
      },
      "CreateGameScoreDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "playedOn": {
            "type": "string"
          },
          "schemaId": {
            "type": "string"
          },
          "scores": {
            "type": "object"
          }
        },
        "required": [
          "gameId",
          "schemaId",
          "scores"
        ]
      },
      "UpdateGameScoreDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "playedOn": {
            "type": "string"
          },
          "schemaId": {
            "type": "string"
          },
          "scores": {
            "type": "object"
          },
          "private": {
            "type": "boolean"
          }
        }
      },
...
      "GameLength": {
        "type": "string",
        "enum": [
          "FILLER",
          "SHORT",
          "MEDIUM",
          "LONG"
        ]
      },
      "GameTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "GameLocationPathDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ]
      },
      "GameLocationDto": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "isGameId": {
            "type": "boolean"
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameLocationPathDto"
            }
          }
        },
        "required": [
          "locationId",
          "isGameId",
          "path"
        ]
      },
      "GameDto": {
        "type": "object",
        "properties": {
          "length": {
            "$ref": "#/components/schemas/GameLength"
          },
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "minPlayers": {
            "type": "number",
            "minimum": 1
          },
          "maxPlayers": {
            "type": "number",
            "minimum": 1
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameTagDto"
            }
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameLocationDto"
            }
          },
          "scoringSchemaIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helperIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdOn": {
            "format": "date-time",
            "type": "string"
          },
          "updatedOn": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "length",
          "id",
          "ownerId",
          "private",
          "name",
          "minPlayers",
          "maxPlayers",
          "tags",
          "locations",
          "scoringSchemaIds",
          "helperIds",
          "createdOn",
          "updatedOn"
        ]
      },
      "CreateGameLocationDto": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string"
          },
          "isGameId": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "locationId"
        ]
      },
      "CreateGameDto": {
        "type": "object",
        "properties": {
          "length": {
            "$ref": "#/components/schemas/GameLength"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "minPlayers": {
            "type": "number",
            "minimum": 1
          },
          "maxPlayers": {
            "type": "number",
            "minimum": 1
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateGameLocationDto"
            }
          },
          "scoringSchemaIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helperIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "length",
          "name",
          "minPlayers",
          "maxPlayers"
        ]
      },
      "UpdateGameDto": {
        "type": "object",
        "properties": {
          "length": {
            "$ref": "#/components/schemas/GameLength"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "minPlayers": {
            "type": "number",
            "minimum": 1
          },
          "maxPlayers": {
            "type": "number",
            "minimum": 1
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateGameLocationDto"
            }
          },
          "scoringSchemaIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helperIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}